refactor/#515 : Command 클래스 알맞은 패키지로 이동
