+++ BE/exceed/src/main/java/com/gaebaljip/exceed/application/port/in/nutritionist/GetDailyAnalysisCommand.java
@@ -1,4 +1,4 @@
-package com.gaebaljip.exceed.adapter.in.nutritionist.request;
+package com.gaebaljip.exceed.application.port.in.nutritionist;
 
 import java.time.LocalDateTime;
